+++ mage_ai/frontend/components/PipelineDetail/AddNewBlocks/v2/constants.tsx
@@ -1,0 +1,9 @@
+import { BlockTypeEnum, ObjectType } from '@interfaces/BlockType';
+import { UIElementConfigurationsType } from '@interfaces/CustomDesignType';
+
+export const ITEMS_MORE: 'items_more' = 'items_more';
+
+export const BUTTON_ITEMS_DEFAULT = [
+  BlockTypeEnum.CUSTOM,
+  BlockTypeEnum.MARKDOWN,
+];
