+++ mage_ai/frontend/interfaces/BlockType.ts
@@ -186,7 +186,7 @@ export interface AnalysisType {
   variable_uuid: string;
 }
 
-enum ObjectType {
+export enum ObjectType {
   BLOCK_FILE = 'block_file',
   CUSTOM_BLOCK_TEMPLATE = 'custom_block_template',
   MAGE_TEMPLATE = 'mage_template',
