+++ mage_ai/frontend/interfaces/CustomDesignType.ts
@@ -9,10 +9,26 @@ interface DesignComponentConfigurationsType {
   media?: MediaType;
 }
 
+export interface UIElementConfigurationsType {
+  items?: string[];
+  items_more?: string[];
+}
+
+interface UIComponentConfigurationsType {
+  add?: UIElementConfigurationsType;
+}
+
+interface ResourceUIComponentConfigurationsType {
+  block?: UIComponentConfigurationsType;
+}
+
+interface PageComponentConfigurationsType {
+  buttons?: ResourceUIComponentConfigurationsType;
+}
+
 interface DesignPageConfigurationsType {
-  edit: {
-    [key: string]: string;
-  };
+  edit?: PageComponentConfigurationsType;
+  list?: PageComponentConfigurationsType;
 }
 
 interface DesignComponentsType {
